Added initial working scripts to repo.
This is the initial addition of the basic configuration reading scripts to the repo.

As of right now (as long as you have the dependencies installed) you should be able to test the configuration reader and writer by running "python3 tadman_main.py". Just pass it a path to a folder holding a legitimate configure file via the CLI.