Added comments to scripts
I was asked by my partner (a week ago) to add comments to these scripts so that he could understand them better. So I added a docstrings to each file, as well as a docstring to each function within the file. Right now they are pretty general, but they may be updated later on.

I also added a line to autotools_config_write which closes the file after it is done making it's list. This should be a little more efficent now.
                                -With love, Ted M.