Added initial GTK interface
After some long and hard work, Ian Swerve and I have formulated a basic interface for Tadman. And so far, it is really great! It contains most of the features I initially had in mind. There are still a few things missing though:
    + A search feature, which does not look solely at the beginning of the line.
    + The ability to add a phrase to the end of an option, especially those which have an equals on their rear, which can be passed multiple different options.
    + Another page with entry boxes for editing standard install directories.
    + The name of the program you are configuring in the information panel.

But all in all, progress is moving along quite nicely. Until next time!
                                ~ Love, Ted M.