Added package name to info panel and cancel button
In this commit, I added an entry box which is preset with the name of the package. This is guesstimated using the name_filter script, which takes in a path, and spits out what is _probably_ the name. Of course, this is not always true, so thus why it is place within an entry box. If the user is not pleased with the current name, they can edit it within the entry box. The default entry right now is only the name, but this is likely to change in the future (with an option to set the version number possibly?).

I also made the small addition of a cancel button. In order for it to make sense, I altered the abilities of the run button. The run button now sets a variable to what is available at the time of hitting run. On the other hand, cancel sets all return values to None to make everything less confusing later on.