Converted README to asciidoc
I have recently become a fan of the asciidoc format, and I am now making
all of my README files and like in the asciidoc format. I also started a
file in the "manual" directory which can be read by Asciidoctor, and
turned into a full man page. It is EXTREMELY convient, and I think everyone
should adopt this format. Thanks.

~ Love, Ted M.