Rewrote main script
I began to notice that the main script was become increasily difficult
to upgrade. So I decided now was a good time to update it. Today I
separated the linking process from the build and install process. The
tadman script now takes arguments including 'build' (which runs a
process similar to the original), 'install' (uses the symlink farm to
'install' the package) and 'uninstall' (basically the opposite of the
'install', so it removes links. Removing directories is still not
supported but will hopefully be safely figured out soon).

Note that for now, the installation process may not function as
intended due to changes within the scripts.