Added fixes to main script and man page
This commit, I made it so that the man page is installed when the scripts are installed. I also added a few fixes, and began the log file creation process. In the future I would like a list option which lists out files that are installed, the date it was built, etc.

~ Love, Ted M.