Update script uniformity and rework main script
With this commit, I revamped the main Tadman script for the most part. It is a little more DRY now and stuff. The variable naming may not be the best, but everything is working just fine now. I have also re-implimented the logger, so now a log is kept whenever a program is installed or uninstalled. Kinda neat. The help message is also pretty now. But that is about it.

~ Love, Ted M.