Update processors pt. 1
In order to better utilize the new curses interface, I have decided to make the options list into an options dictionary. I have essentially rewritten the autotools processor, but I have yet to do so for CMake. So that will be done soon. Also, the curses interface is a little more spiffy. It is really close to production-ready.

~ Love, Ted M.