Update interfaces and processors
So... Now the interfaces and processors both work in terms of dictionaries, (or ordered dictionaries to be more specific). Now both interfaces output the very same output format, so that is good news. The next step is to update the configuration output script, and then add some changes to the tadman script. Then everything will be fine and dandy :)

~ Love, Ted M.