Add scrolling to curses interface
The curses interface now has the ability to scroll! So now it can list more than 15 options in a stylish manner. Also did some minor refactoring of class names and the main script works once again. Stil question as to whether the main script will be rewritten, but we will see in the next few days.

~ Love, Ted M.