Added 'list' argument
Tadman now has the ability to list out either all the packages you have compile in your package directory, or the files installed by a package. This change has been made in the help message, as well as the manpage.

I also move a few functions around. But it is all fine and dandy.