Add new window to curses interface
I have been working on adding another window to the curses interface.
This is where the installation path flags will be changed, so things
like the mandir and such can be edited within Tadman. By having this
ability, Tadman is that much closer to simplifying the entire source
code building process. So great news in fact.

~ Love, Ted M.