Update commenting in lib scripts
Some of the scripts that were edited recently for the install interface
were updated to be more accurate and complete. They are still not
perfect though...

Anyways, future plans include a help message box that will overlap the
current menu, and the ability to put 'YES' or 'NO' for CMake optional
features. I am hoping to complete these moderately soon.

~ Love, Ted M.