Add CMake 'ON' and 'OFF' support and tiny verison bump
I realized earlier today that CMake optional features can be turned on
and off. So I decided to incorporate this quickly before I merge this
into the master branch. By pressing 'y', the feature will be turned on,
'n' will turn it off and enter will clear the value.

I also threw in a quick little version bump to '0.1.2'. Now just for a
quick little merge.

~ Love, Ted M.