Add help message loop
Now all you have to do in the interface to see all of the keys available
to the user is enter a '?'. And it works splendidly!

I have come across the fact the HOME and END keys do not work. Hopefully
by the end of the day I can get them up and running and tidy everything
up with comments. After that feature is complete, there will likely be
another merge and small version bump to 1.4 or something. Then try to
pick up some more publicity or something.

~ Love, Ted M.