Fix tiny bug and move documents to reStructuredText
In this commit, I fixed a tiny line wrapping bug (Which I think I fixed
before but never pushed).

In other news, since this is a Python project, I have decided to move
over to reStructuredText since the state of Python's Asciidoc is quite
dismal right now. I think it may be for the best. One of the biggest
changes is the manpage, which can be easily installed using docutils and
the `python setup.py install_manpage` as root user. Of course, you need
docutils installed. But other than that, it is easier than ever before!

~ Love, Ted M.